refactor: DatabaseService #29
